Add support for loading multiple game save versions
