+++ backend/engine/extraction/__init__.py
@@ -1,7 +1,7 @@
 from . import extractors
 from . import postprocessors
 from . import isolation_layer
-from . import util
+from .isolation import util
 
 extractor_list = extractors.build_extractor_list(isolation_layer)
 postprocessor_list = postprocessors.build_postprocessor_list(isolation_layer)
