+++ backend/engine/extraction/isolation/__init__.py
@@ -1,0 +1,3 @@
+from . import v1
+
+VERSIONS = [v1]
