+++ backend/engine/extraction/isolation/util.py
